Added basic authentication to the sync routes. Modified the device schema to add a key and added_by field. Modified the device middleware to validate the device key. Added check existing reports for an old migration so that it doesn't cause an error.

+++ database/migrations/2019_01_18_160537_add_study_id_form_id_to_report.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Models\Report;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
@@ -13,6 +14,11 @@ class AddStudyIdFormIdToReport extends Migration
      */
     public function up()
     {
+        $reportCount = Report::count();
+        if ($reportCount > 0) {
+            throw new Exception("The reports table should be truncated before running this migration");
+        }
+
         Schema::table('report', function (Blueprint $table) {
             $table->dropColumn('report_id');
